# internal/docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  request.UpdateTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.MessageListResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/response.MessageResponse'
        type: array
    type: object
  response.MessageResponse:
    properties:
      content:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      pagination:
        $ref: '#/definitions/response.PaginationInfo'
      success:
        type: boolean
    type: object
  response.PaginationInfo:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      success:
        type: boolean
    type: object
...
  response.TagListResponse:
    properties:
      tags:
        items:
          $ref: '#/definitions/response.TagResponse'
        type: array
    type: object
  response.TagResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@chatassistant.com
    name: Chat Assistant Team
  description: A RESTful API for the Chat Assistant Backend service. This API provides
    endpoints for managing users, conversations, and messages.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Chat Assistant Backend API
  version: 1.0.0
paths:
...
  /api/v1/conversations/{id}/messages:
    get:
      consumes:
      - application/json
      description: Retrieve all messages in a specific conversation with pagination
      parameters:
      - description: Conversation ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Messages list
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.MessageListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Conversation Messages
      tags:
      - Conversations
...
  /api/v1/messages:
    get:
      consumes:
      - application/json
      description: Retrieve messages list with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Messages list
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.MessageListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Messages
      tags:
      - Messages
  /api/v1/messages/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific message by ID
      parameters:
      - description: Message ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Message not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete Message
      tags:
      - Messages
    get:
      consumes:
      - application/json
      description: Retrieve a specific message by ID
      parameters:
      - description: Message ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message details
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MessageResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Message not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Message
      tags:
      - Messages
...
  /api/v1/tags:
    get:
      consumes:
      - application/json
      description: Retrieve all tags
      produces:
      - application/json
      responses:
        "200":
          description: Tags list
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TagListResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      description: Create a new tag
      parameters:
      - description: Tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/request.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tag created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TagResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Tag name already exists
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create Tag
      tags:
      - Tags
  /api/v1/tags/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific tag by ID
      parameters:
      - description: Tag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Tag not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete Tag
      tags:
      - Tags
    get:
      consumes:
      - application/json
      description: Retrieve a specific tag by ID
      parameters:
      - description: Tag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag details
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TagResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Tag not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Tag
      tags:
      - Tags
    put:
      consumes:
      - application/json
      description: Update an existing tag
      parameters:
      - description: Tag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tag updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TagResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Tag not found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Tag name already exists
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update Tag
      tags:
      - Tags
...
swagger: "2.0"